+++ tests/Factory/User/Mailer/WelcomeEmailFactoryTest.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Silverback\ApiComponentsBundle\Tests\Factory\Mailer\User;
+namespace Silverback\ApiComponentsBundle\Tests\Factory\User\Mailer;
 
 use Silverback\ApiComponentsBundle\Entity\User\AbstractUser;
 use Silverback\ApiComponentsBundle\Factory\User\Mailer\WelcomeEmailFactory;
@@ -57,9 +57,8 @@ class WelcomeEmailFactoryTest extends AbstractFinalEmailFactoryTest
         };
         $user
             ->setUsername('username')
-            ->setEmailAddress('email@address.com')
-            ->setNewEmailVerificationToken('token');
-
+            ->setEmailAddress('email@address.com');
+        $user->plainNewEmailVerificationToken = 'token';
         $factory = new WelcomeEmailFactory($this->containerInterfaceMock, $this->eventDispatcherMock, 'subject', true, '/default-path');
 
         $this->assertCommonMockMethodsCalled(true);
